refactor: move 패키지 생성에따른 depth 변경
