+++ src/main/java/step3/strategy/move/MoveStrategy.java
@@ -1,4 +1,4 @@
-package step3.strategy;
+package step3.strategy.move;
 
 @FunctionalInterface
 public interface MoveStrategy {
